Downgrade requirements for WASM dependencies
WebAssembly support was added in #38 targeting the current versions
versions of wasm-bindgen and js-sys at the time, which is
reasonable. Unfortunately the whoami crate, in trying to support
older Rust versions, has reduced the upper version of wasm-bindgen
that it supports, see https://github.com/ardaku/whoami/issues/43.

This PR thus reduces the minimum version required for wasm-bindgen.
I used `cargo +nightly update -Z minimal-versions` with
`cargo build --target wasm32-unknown-unknown` to find the earliest
version that compiles. We could theoretically set up checks in
CI to verify this on an ongoing basis, but this might be overkill
for now.