Replace `winapi` dependency with `windows-sys`
[Some time ago] I spoke in favor of using [`winapi`] instead of
[`windows-sys`], because it had fewer msrv bumps, and because it was
used by e.g. `tokio`, and `chrono`, so even if it was not being
developed for anymore, it still had a ton of users who would hopefully
notice any bugs.

By now `tokio` has switched to `windows-sys`, and `chrono` will too in
the next release. I guess it's time then to make the switch, too.

[Some time ago]: https://github.com/strawlab/iana-time-zone/pull/35
[`winapi`]: https://crates.io/crates/winapi
[`windows-sys`]: https://crates.io/crates/windows-sys