Implement `windows` backend without `windows-sys`
This PR removes the `windows-sys` and `winapi` dependency completely.
Instead, the relevant API was extracted from `windows-sys`, which in
turn is generated from Windows's header files. The only Windows specific
depency now is `windows-targets`, which was a transitive dependency
before. It contains the proper mappings between a function name, its
mangled name, and its `.dll`.

The rationale for this change is that the previously used `winapi` crate
is unmaintained for the last three years, even though it contains some
(minor) errors.

The alternative to `winapi`, `windows-sys` is still fast moving. In its
(at the time of writing this PR) newest release 0.45.0 it removed all
WinRT functions, so it arguably does not any function we need to
implement our library anymore.