Use C ABI for haiku platform support native bindings
The haiku platform support crate uses `cxx` and `cxx-build` for its
native bindings to the `be` SDK, however `cxx` is a heavy dependency
which is not necessary for interfacing with a C++ native library.

Replace `cxx-build` with `cc` set to build for C++ mode. Instead of
exporting a C++ ABI interace and bridging with `cxx`, use the C ABI
(extern "C" in both C++ and Rust).

This introduces one line of `unsafe` which is documented with SAFETY
comments (and is not meaningfully different from the C++ bridge). The
C++ source file is updated to use the new function arguments.

As an opportunistic refactor, add a `#[must_use]` attribute to
`get_timezone` in the haiku crate.