Allow version range for `windows` dependency
`iana-time-zone` works with any `windows` version since `0.30.0`. We
don't have to care which version the user actually uses. We only need
to make sure that the selected version is not too old. Let `cargo`
decide everything else.

The newest version of `windows`, `0.51.0`, has an msrv of `1.56`. If a
user of `iana-time-zone` is not able or willing to update their rust
installation, and want to continue using rustc `1.48`, they can select
an older version of the `windows` crate, and still receive any updates
`iana-time-zone` might have in the future.