Bump `windows-core` to `0.55`
Note that `windows 0.55` was not released, and any implementation in
the wild depending on `windows` will get `windows-core 0.54` in their
dependency tree for obvious compatibility reasons, and would get a
duplicate if `iana-time-zone` decides to switch to o`windows-core 0.55`
prematurely. This PR depends on #131 and can be updated and merged
whenever there is a compelling reason to bump further.

This requires a strange change to remove a "comment" from `bindings.txt`:

         Running `api_gen/target/debug/api_gen`
    Error: Error { message: "unused filters\n  Dependencies\n  above\n  the\n  of\n  #", path: "", span: None }